chore: move input template test to a separate file
